codemaster/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
SQLAlchemy==2.0.36
alembic==1.13.2
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
pyjwt==2.9.0
aiosqlite==0.20.0
